pyproject.toml
[project]
name = "blog-comment"
version = "0.1.0"
description = "A blog that allows visitors to post comments"
authors = [
    {name = "LH_cat, Raydev"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-wtf (>=1.2.2,<2.0.0)",
    "flask-login (>=0.6.3,<0.7.0)",
    "flask-uploads (>=0.2.1,<0.3.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.9.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "pytest (>=9.0.0,<10.0.0)",
    "pytest-flask (>=1.3.0,<2.0.0)",
    "ruff (>=0.14.5,<0.15.0)",
    "pyright (>=1.1.408,<2.0.0)"
]

[tool.ruff]
line-length = 320

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "Q", "T"]
